requirements.txt
accelerate==1.11.0
altair==5.5.0
asttokens==3.0.0
attrdict==2.0.1
attrs==25.4.0
bitsandbytes==0.48.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
deepseek-vl==1.0.0
einops==0.8.1
executing==2.2.1
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython-pygments-lexers==1.1.1
ipython==9.7.0
jedi==0.19.2
jinja2==3.1.6
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
jupyter-client==8.6.3
jupyter-core==5.9.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
pip==24.0
platformdirs==4.5.0
prompt-toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
pure-eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
pygments==2.19.2
pyngrok==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scipy==1.16.3
sentencepiece==0.2.1
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
timm==1.0.22
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
wheel==0.45.1
